Test for the new function
